python/ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-genai>=1.67.0",
]
